+++ src/main/java/umc/Entity/alert/Alert.java
@@ -1,0 +1,33 @@
+package umc.Entity.alert;
+
+import jakarta.persistence.*;
+import lombok.*;
+import umc.Entity.BaseTimeEntity;
+import umc.Entity.user.User;
+
+@Builder
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Table(name = "alert")
+public class Alert extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false, unique = true)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+
+    @Column(name = "title", nullable = false)
+    private String title;
+
+    @Column(name = "content", nullable = false)
+    private String content;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "alert_category_id", nullable = false)
+    private AlertCategory alertCategory;
+}
