+++ src/main/java/umc/Entity/alert/AlertCategory.java
@@ -1,4 +1,4 @@
-package umc.domain.alert;
+package umc.Entity.alert;
 
 import jakarta.persistence.*;
 import lombok.*;
