+++ src/main/java/umc/Entity/inquiry/InquiryAnswer.java
@@ -1,33 +1,27 @@
-package umc.domain.alert;
+package umc.Entity.inquiry;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.domain.BaseTimeEntity;
-import umc.domain.user.User;
 
 @Builder
 @Entity
-@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-@Table(name = "alert")
-public class Alert extends BaseTimeEntity {
+@Table(name = "inquiry_answer")
+@Getter
+public class InquiryAnswer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, unique = true)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "title", nullable = false)
+    @Column(name ="title", nullable = false)
     private String title;
 
     @Column(name = "content", nullable = false)
     private String content;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "category_id", nullable = false)
-    private AlertCategory alertCategory;
+    @JoinColumn(name = "inquiry_id", nullable = false)
+    private Inquiry inquiry;
 }
