+++ src/main/java/umc/Entity/mission/MissionState.java
@@ -1,4 +1,4 @@
-package umc.domain.mission;
+package umc.Entity.mission;
 
 import lombok.Getter;
