+++ src/main/java/umc/Entity/store/StoreImage.java
@@ -1,4 +1,4 @@
-package umc.domain.store;
+package umc.Entity.store;
 
 
 import jakarta.persistence.*;
