--- src/main/java/umc/domain/inquiry/InquiryAnswer.java
@@ -1,27 +1,0 @@
-package umc.domain.inquiry;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Builder
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Table(name = "inquiry_answer")
-@Getter
-public class InquiryAnswer {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false, unique = true)
-    private Long id;
-
-    @Column(name ="title", nullable = false)
-    private String title;
-
-    @Column(name = "content", nullable = false)
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "inquiry_id", nullable = false)
-    private Inquiry inquiry;
-}
